+++ lib/omq/options.rb
@@ -24,7 +24,7 @@ module OMQ
       @heartbeat_interval    = nil   # seconds, nil = disabled
       @heartbeat_ttl         = nil   # seconds, nil = use heartbeat_interval
       @heartbeat_timeout     = nil   # seconds, nil = use heartbeat_interval
-      @max_message_size      = 1 << 20  # bytes (1 MiB default)
+      @max_message_size      = nil      # bytes, nil = unlimited
       @conflate              = false
       @sndbuf                = nil   # bytes, nil = OS default
       @rcvbuf                = nil   # bytes, nil = OS default